_mload -> mload - gas efficiency
Initially _mload was created to add a check - if memsize < pointer value, revert with error code.
But this is highly ineffcient. ~450 gas costs shaved off by removing it for a simple add_ operation.

Also, the check should not be needed at this low level. This is the unsafe, untyped, fast core.